fix: ForgeStateExtractor - rangeSq type (Double) for FOV-nearest distance compare
range is Float; distSq is Double. Coerce rangeSq to Double to fix :adapter:forge compile.

+++ SwitchLite/adapter/forge/v1_8_9/src/main/kotlin/io/switchlite/adapter/forge/v1_8_9/ForgeStateExtractor.kt
@@ -385,7 +385,7 @@ object ForgeStateExtractor : IStateExtractor {
         val playerPitch = MappingContext.getFieldValue(player, "forge:player_rotationPitch") as? Float ?: 0f
 
         val loadedList = MappingContext.getFieldValue(world, "forge:world_loadedEntityList") as? List<*> ?: return null
-        val rangeSq = range * range
+        val rangeSq = (range.toDouble() * range.toDouble())
 
         var nearestId: Int? = null
         var nearestDistSq = Double.MAX_VALUE
